+++ src/client/index.js
@@ -7,17 +7,15 @@ import intl from 'intl';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import api from 'api/index';
 import App from './App';
 
 // global styles
-import './style.scss';
+import './style/style.scss';
 
 // apply polyfill
 if (!window.Intl) {
   window.Intl = intl;
 }
 
-api.setEndpoint('/api');
 
 ReactDOM.render(<App />, document.getElementById('app'));