+++ src/client/router.js
@@ -1,16 +1,13 @@
 import React from 'react';
-import {BrowserRouter, Route, Switch} from 'react-router-dom';
+import {HashRouter, Route, Switch} from 'react-router-dom';
 
-import AppLayout from 'containers/AppLayout';
-import LoginForm from 'containers/LoginForm';
-
+import AppLayout from './containers/AppLayout';
 export default function() {
   return (
-    <BrowserRouter>
+    <HashRouter>
       <Switch>
-        <Route path="/login" name="login" component={LoginForm} />
-        <Route path="/" name="home" component={AppLayout} />
+        <Route path="/" exact name="home" component={AppLayout} />
       </Switch>
-    </BrowserRouter>
+    </HashRouter>
   );
 }