+++ src/server/index.js
@@ -1,11 +1,18 @@
-const http = require('http');
+const https = require('https');
 const express = require('express');
 const bodyParser = require('body-parser');
 const setupApiRoutes = require('./middlewares/api');
 const logger = require('./logger');
+const compression = require('compression');
+const fs = require('fs');
 
+const http = require('http');
 process.env.NODE_ENV = process.env.NODE_ENV || 'development';
 process.env.HTTP_PORT = process.env.HTTP_PORT || 3000;
+const app = express();
+const server = http.Server(app);
+
+//const switchboard = require('rtc-switchboard')(server);
 
 function onUnhandledError(err) {
   try {
@@ -23,18 +30,17 @@ process.on('uncaughtException', onUnhandledError);
 const setupAppRoutes =
   process.env.NODE_ENV === 'development' ? require('./middlewares/development') : require('./middlewares/production');
 
-const app = express();
-
 app.set('env', process.env.NODE_ENV);
 logger.info(`Application env: ${process.env.NODE_ENV}`);
 
 app.use(logger.expressMiddleware);
 app.use(bodyParser.json());
-
+app.use(compression());
+//app.get('/rtc.io/primus.js', switchboard.library());
 // application routes
 setupApiRoutes(app);
 setupAppRoutes(app);
 
-http.createServer(app).listen(process.env.HTTP_PORT, () => {
+server.listen(process.env.HTTP_PORT, () => {
   logger.info(`HTTP server is now running on http://localhost:${process.env.HTTP_PORT}`);
 });